# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Knowledge Melting Pot API
  version: 1.0.0
  description: Versioned API for auditable knowledge extraction and publication.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /subscenes/{subSceneId}/assets:
    parameters:
      - $ref: '#/components/parameters/SubSceneId'
    get:
      summary: List the latest asset for each type of a sub-scene
      operationId: listAssets
      responses:
        '200':
          description: Latest assets; BLOCKED means the generation prerequisite is missing.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Asset'
        '404':
          $ref: '#/components/responses/Problem'
  /assets/{assetId}/download:
    parameters:
      - name: assetId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Redirect to a short-lived presigned download URL for an immutable asset bundle
      description: Only READY assets can be downloaded; the redirect URL expires in minutes and never exposes object-storage credentials.
      operationId: downloadAsset
      responses:
        '302':
          description: Redirect to a short-lived presigned URL.
        '200':
          description: Resource body after following the redirect (the immutable asset bundle).
        '404':
          $ref: '#/components/responses/Problem'
  /subscenes/{subSceneId}/asset-generation-jobs:
    post:
      summary: Generate one or more knowledge assets
      operationId: generateAssets
      parameters:
        - $ref: '#/components/parameters/SubSceneId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateAssetsRequest'
      responses:
        '202':
          $ref: '#/components/responses/AcceptedJob'
        '400':
          $ref: '#/components/responses/Problem'
...
  /knowledge-documents/{documentId}:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    get:
      summary: Get the current knowledge document revision
      operationId: getKnowledgeDocument
      responses:
        '200':
          description: Current document revision.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeDocument'
        '404':
          $ref: '#/components/responses/Problem'
    put:
      summary: Validate and save a new document revision
      operationId: saveKnowledgeDocument
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveDocumentRequest'
      responses:
        '200':
          description: Newly created immutable revision.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeDocument'
        '412':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /knowledge-documents/{documentId}/revisions:
    get:
      summary: List immutable document revisions
      operationId: listDocumentRevisions
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: Immutable revision history.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentRevisionSummary'
        '404':
          $ref: '#/components/responses/Problem'
  /knowledge-documents/{documentId}/alignment-jobs:
    post:
      summary: Start an AI alignment proposal job
      operationId: startAlignment
      parameters:
        - $ref: '#/components/parameters/DocumentId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartAlignmentRequest'
      responses:
        '202':
          $ref: '#/components/responses/AcceptedJob'
        '400':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '412':
          $ref: '#/components/responses/Problem'
  /knowledge-documents/{documentId}/alignment-proposals:
    get:
      summary: List immutable alignment proposals for a knowledge document
      operationId: listAlignmentProposals
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: Proposals ordered newest first, including typed diff and adoption state.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AlignmentProposal' }
        '404':
          $ref: '#/components/responses/Problem'
  /alignment-proposals/{proposalId}:
    get:
      summary: Get an immutable alignment proposal
      operationId: getAlignmentProposal
      parameters:
        - name: proposalId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Immutable proposal plus separately recorded adoption metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlignmentProposal'
        '404':
          $ref: '#/components/responses/Problem'
  /alignment-proposals/{proposalId}/adopt:
    post:
      summary: Adopt a proposal against its base revision
      operationId: adoptAlignmentProposal
      parameters:
        - name: proposalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Proposal applied as a new document revision.
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeDocument'
        '409':
          $ref: '#/components/responses/Problem'
        '412':
          $ref: '#/components/responses/Problem'
        '428':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: KMP_SESSION
  parameters:
...
    SubSceneId:
      name: subSceneId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    DocumentId:
      name: documentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
  responses:
    AcceptedJob:
      description: Command accepted as a durable job.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AcceptedJob'
    Problem:
      description: RFC 7807 problem response.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    AssetType:
      type: string
      enum: [RULE_CATALOG, DECISION_FLOW, SKILL_PACKAGE, QA_PAIRS, EVALUATION_SET]
    AssetStatus:
      type: string
      enum: [PENDING, GENERATING, READY, FAILED, SUPERSEDED, BLOCKED]
    Asset:
      type: object
      additionalProperties: false
      required: [id, subSceneId, type, version, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        subSceneId: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/AssetType' }
        version: { type: integer, minimum: 1 }
        status: { $ref: '#/components/schemas/AssetStatus' }
        documentRevisionId: { type: [string, 'null'], format: uuid }
        objectKey: { type: string }
        checksum: { type: string }
        failureReason: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    GenerateAssetsRequest:
      type: object
      additionalProperties: false
      required: [documentRevisionId, types]
      properties:
        documentRevisionId: { type: string, format: uuid }
        types:
          type: array
          minItems: 1
          uniqueItems: true
          items: { $ref: '#/components/schemas/AssetType' }
    StartAlignmentRequest:
      type: object
      additionalProperties: false
      description: Schedules alignment only; document bodies and arbitrary parameters are not accepted.
      required: [baseRevisionId, action, regulatoryMaterialIds]
      properties:
        baseRevisionId: { type: string, format: uuid }
        action:
          type: string
          enum: [CONSISTENCY, REGULATORY, GAP_ANALYSIS, REWRITE]
        regulatoryMaterialIds:
          type: array
          maxItems: 100
          uniqueItems: true
          description: Required and non-empty only for REGULATORY; every item must be regulatory and non-HOLDOUT.
          items: { type: string, format: uuid }
    AlignmentProposal:
      type: object
      additionalProperties: false
      required: [id, documentId, baseRevisionId, baseEtag, action, status, structuredPatch,
        reason, sourceRefs, regulatoryMaterialIds, createdBy, createdAt]
      properties:
        id: { type: string, format: uuid }
        documentId: { type: string, format: uuid }
        baseRevisionId: { type: string, format: uuid }
        baseEtag: { type: string }
        action: { type: string, enum: [CONSISTENCY, REGULATORY, GAP_ANALYSIS, REWRITE] }
        status: { type: string, enum: [READY, ADOPTED] }
        structuredPatch: { $ref: '#/components/schemas/KnowledgePatch' }
        reason: { type: string }
        sourceRefs:
          type: array
          items: { $ref: '#/components/schemas/SourceRef' }
        regulatoryMaterialIds:
          type: array
          items: { type: string, format: uuid }
        createdBy: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        adoptedRevisionId: { type: [string, 'null'], format: uuid }
        adoptedBy: { type: [string, 'null'], format: uuid }
        adoptedAt: { type: [string, 'null'], format: date-time }
    KnowledgePatch:
      type: object
      additionalProperties: false
      required: [operation, replacement, diff]
      properties:
        operation: { type: string, const: replaceKnowledgeIr }
        replacement: { $ref: '#/components/schemas/KnowledgeIr' }
        diff: { $ref: '#/components/schemas/KnowledgeDiff' }
    KnowledgeDiff:
      type: object
      additionalProperties: false
      required: [addedRuleIds, removedRuleIds, changedRuleIds, addedFlowIds, removedFlowIds,
        changedFlowIds, sourceRefDelta]
      properties:
        addedRuleIds: { type: array, items: { type: string } }
        removedRuleIds: { type: array, items: { type: string } }
        changedRuleIds: { type: array, items: { type: string } }
        addedFlowIds: { type: array, items: { type: string } }
        removedFlowIds: { type: array, items: { type: string } }
        changedFlowIds: { type: array, items: { type: string } }
        sourceRefDelta: { type: integer }
    AcceptedJob:
      type: object
      required: [jobId, status, statusUrl, eventsUrl]
      properties:
        jobId: { type: string, format: uuid }
        status:
          type: string
          enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, CANCELLED]
        statusUrl: { type: string }
        eventsUrl: { type: string }
        resourceType: { type: [string, 'null'] }
        resourceId: { type: [string, 'null'], format: uuid }
...
    SourceRef:
      type: object
      additionalProperties: false
      description: >
        A source reference derived exclusively from the persisted chunk locator
        and material identity. Client input never contributes to these fields.
      required: [code, materialId, materialSha256, chunkId, locatorType, excerptHash]
      properties:
        code: { type: string, pattern: '^SRC-[A-Za-z0-9_-]{1,100}$' }
        materialId: { type: string, format: uuid }
        materialSha256: { type: string }
        chunkId: { type: string, format: uuid }
        locatorType:
          type: string
          enum: [PDF_PAGE_PARAGRAPH, DOCX_PARAGRAPH, DOCX_TABLE_CELL, XLSX_RANGE, TXT_LINES]
        page: { type: [integer, 'null'], minimum: 1 }
        paragraph: { type: [integer, 'null'], minimum: 0 }
        table: { type: [integer, 'null'], minimum: 0 }
        sheet: { type: [string, 'null'] }
        rowStart: { type: [integer, 'null'], minimum: 0 }
        rowEnd: { type: [integer, 'null'], minimum: 0 }
        colStart: { type: [integer, 'null'], minimum: 0 }
        colEnd: { type: [integer, 'null'], minimum: 0 }
        lineStart: { type: [integer, 'null'], minimum: 0 }
        lineEnd: { type: [integer, 'null'], minimum: 0 }
        excerptHash: { type: string }
    KnowledgeIr:
      type: object
      additionalProperties: false
      required: [schemaVersion, metadata, rules, flows, conflicts, gaps, sourceRefs]
      properties:
        schemaVersion: { type: string, const: knowledge-ir/v1 }
        metadata:
          type: object
          additionalProperties: false
          required: [documentId, subSceneId, roundId, inputHash]
          properties:
            documentId: { type: string, format: uuid }
            subSceneId: { type: string, format: uuid }
            roundId: { type: string, format: uuid }
            inputHash: { type: string, pattern: '^[a-f0-9]{64}$' }
        rules:
          type: array
          items: { $ref: '#/components/schemas/KnowledgeRule' }
        flows:
          type: array
          items: { $ref: '#/components/schemas/KnowledgeFlow' }
        conflicts:
          type: array
          items: { $ref: '#/components/schemas/KnowledgeConflict' }
        gaps:
          type: array
          items: { $ref: '#/components/schemas/KnowledgeGap' }
        sourceRefs:
          type: array
          items: { $ref: '#/components/schemas/SourceRef' }
    KnowledgeRule:
      type: object
      additionalProperties: false
      required: [id, title, condition, conclusion, priority, exceptions, sourceRefs]
      properties:
        id: { type: string, pattern: '^R-[a-f0-9]{16}$' }
        title: { type: string }
        condition: { type: string }
        conclusion: { type: string }
        priority: { type: integer, minimum: 0, maximum: 1000 }
        exceptions: { type: array, items: { type: string } }
        sourceRefs:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string, pattern: '^SRC-[A-Za-z0-9_-]{1,100}$' }
    KnowledgeFlow:
      type: object
      additionalProperties: false
      required: [id, name, nodes, edges]
      properties:
        id: { type: string }
        name: { type: string }
        nodes:
          type: array
          items: { $ref: '#/components/schemas/KnowledgeFlowNode' }
        edges:
          type: array
          items: { $ref: '#/components/schemas/KnowledgeFlowEdge' }
    KnowledgeFlowNode:
      type: object
      additionalProperties: false
      required: [id, label, critical, sourceRefs]
      properties:
        id: { type: string }
        label: { type: string }
        critical: { type: boolean }
        sourceRefs:
          type: array
          uniqueItems: true
          items: { type: string, pattern: '^SRC-[A-Za-z0-9_-]{1,100}$' }
    KnowledgeFlowEdge:
      type: object
      additionalProperties: false
      required: [id, source, target, label]
      properties:
        id: { type: string }
        source: { type: string }
        target: { type: string }
        label: { type: string }
    KnowledgeConflict:
      type: object
      additionalProperties: false
      required: [id, description, sourceRefs]
      properties:
        id: { type: string }
        description: { type: string }
        sourceRefs:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string, pattern: '^SRC-[A-Za-z0-9_-]{1,100}$' }
    KnowledgeGap:
      type: object
      additionalProperties: false
      required: [id, description]
      properties:
        id: { type: string }
        description: { type: string }
    KnowledgeDocument:
      type: object
      required: [id, subSceneId, revisionId, revisionNumber, contentMd, contentHash, sourceRefs, etag]
      properties:
        id: { type: string, format: uuid }
        subSceneId: { type: string, format: uuid }
        revisionId: { type: string, format: uuid }
        revisionNumber: { type: integer, minimum: 1 }
        contentMd: { type: string }
        contentHash: { type: string }
        finalized: { type: boolean }
        sourceRefs:
          type: array
          items: { $ref: '#/components/schemas/SourceRef' }
        etag: { type: string }
    SaveDocumentRequest:
      type: object
      additionalProperties: false
      required: [contentMd, finalize]
      properties:
        subSceneId:
          type: [string, 'null']
          format: uuid
          description: Required only when creating the first revision.
        contentMd: { type: string, minLength: 1 }
        revisionNote: { type: string, maxLength: 500 }
        finalize: { type: boolean, default: false }
    DocumentRevisionSummary:
      type: object
      additionalProperties: false
      required: [id, revisionNumber, contentHash, createdBy, finalized, createdAt]
      properties:
        id: { type: string, format: uuid }
        revisionNumber: { type: integer }
        contentHash: { type: string }
        note: { type: [string, 'null'] }
        createdBy: { type: string, format: uuid }
        finalized: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    Problem:
      type: object
      required: [type, title, status, code, traceId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        code: { type: string }
        traceId: { type: string }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }